apps/realty-main/requirements.txt
# Minimal Railway requirements for Dubai MVP v3.0
# Version: 3.0 - Ultra-simplified for Railway
# Build: 2025-08-25 - Force Railway to rebuild
django>=4.2
djangorestframework>=3.15
django-cors-headers>=4.7
gunicorn>=21.2
environs[django]>=9.5
psycopg[binary]>=3.2
whitenoise[brotli]>=6.6
djangorestframework-simplejwt>=5.3
prometheus-client>=0.21
